dvfs: Extract inode_operations to separate header
